目的地を持たせて，その方向へ進めればかなり終わる． 進めればかなり終わる

+++ Lost/Lost/Value.h
@@ -73,8 +73,8 @@ const int MADAM_SERCH_HEIGHT = 90;
 
 /*--------’·’j--------*/
 const int SON_RANGE = 60;
-const int SON_FULL_SPEED = 35;
-const int SON_HALF_SPEED = 20;
+const double SON_FULL_SPEED = 35.0;
+const double SON_HALF_SPEED = 20.0;
 const int SON_SERCH_WIDTH = 1100;
 const int SON_SERCH_HEIGHT = 150;
 /*--------------------*/
